+++ utils/defaultsettings.gd
@@ -11,6 +11,7 @@ var overlay_dimundim = true
 var overlay_dimdownopacity = 0.4
 var overlay_dimdownafter = 30
 var overlay_undimchime = true
+var overlay_chimedevice = "Default"
 
 # Overlay position / rotation
 var overlay_size = 1.5